refactor: move components to the separate folders
