+++ src/components/form-elements/text-inputs/InputPassword.vue
