+++ src/components/forms/RegistrationForm.vue
@@ -87,18 +87,18 @@
 
 <script lang="ts">
 import useValidate from '@vuelidate/core'
-import InputName from '@/components/form-elements/InputName.vue'
-import InputLastName from '@/components/form-elements/InputLastName.vue'
-import InputDate from '@/components/form-elements/InputDate.vue'
-import InputEmail from '@/components/form-elements/InputEmail.vue'
-import InputPassword from '@/components/form-elements/InputPassword.vue'
-import InputStreet from '@/components/form-elements/InputStreet.vue'
-import InputCity from '@/components/form-elements/InputCity.vue'
-import InputPotscode from '@/components/form-elements/InputPotscode.vue'
-import InputCountry from '@/components/form-elements/InputCountry.vue'
-import InputBuilding from '@/components/form-elements/InputBuilding.vue'
-import InputApartment from '@/components/form-elements/InputApartment.vue'
-import MyCheckbox from '@/components/form-elements/radio/MyCheckbox.vue'
+import InputName from '@/components/form-elements/text-inputs/InputName.vue'
+import InputLastName from '@/components/form-elements/text-inputs/InputLastName.vue'
+import InputDate from '@/components/form-elements/text-inputs/InputDate.vue'
+import InputEmail from '@/components/form-elements/text-inputs/InputEmail.vue'
+import InputPassword from '@/components/form-elements/text-inputs/InputPassword.vue'
+import InputStreet from '@/components/form-elements/text-inputs/InputStreet.vue'
+import InputCity from '@/components/form-elements/text-inputs/InputCity.vue'
+import InputPotscode from '@/components/form-elements/text-inputs/InputPotscode.vue'
+import InputCountry from '@/components/form-elements/text-inputs/InputCountry.vue'
+import InputBuilding from '@/components/form-elements/text-inputs/InputBuilding.vue'
+import InputApartment from '@/components/form-elements/text-inputs/InputApartment.vue'
+import MyCheckbox from '@/components/form-elements/checkboxes/MyCheckbox.vue'
 import { createCustomerDraft, createAddressDraft } from '@/helpers/createDataSamples'
 import type { IAddressConfiguration, IAddressDraft } from '@/types/customer-types'
 import { useApiRootStore } from '@/stores/ApiRoot'
